+++ packages/zephyr/src/Forms/Label.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import VisuallyHidden from '@reach/visually-hidden';
 import { BoxProps } from '../Box';
 import { Text } from '../Text';
@@ -33,6 +32,9 @@ export interface LabelProps extends Pick<BoxProps, 'children' | 'className' | 't
   size?: 12 | 14;
 }
 
+/**
+ * @deprecated This component is no longer supported and will soon be removed. It will be replaced with a `LabelPrimitive` component.
+ */
 export const Label = ({
   children,
   className,