+++ packages/zephyr/src/RadixMenu/abstractions/RadixDropdownMenu.tsx
@@ -1,6 +1,6 @@
 import { AnimatePresence, Variant } from 'framer-motion';
 import { noop } from 'lodash';
-import React, { ReactNode, useState, useCallback, useMemo, memo } from 'react';
+import { ReactNode, useState, useCallback, useMemo, memo } from 'react';
 import { motion, useReducedMotion } from 'framer-motion';
 import { Trigger, Content, Root } from '@radix-ui/react-dropdown-menu';
 import { Slot } from '@radix-ui/react-slot';
@@ -124,7 +124,14 @@ export const RadixDropdownMenu = memo(
     );
 
     return (
-      <Root onOpenChange={handleChange} modal={false}>
+      <Root
+        onOpenChange={handleChange}
+        modal={false}
+        dir="ltr"
+        // `dir` must be provided to in order to prevent Radix from kicking a
+        // looping useDirection hook which clogs the JS thread on browser.
+        // @see https://github.com/radix-ui/primitives/issues/749
+      >
         {/* This overlay is conditional rendered based on if `hasOverlay` prop is passed and it'll apply a transparent div over the entire screen to prevent the user from being able to right click on anything else while the menu is opened. */}
         {hasOverlay && (
           <Box