+++ packages/zephyr/stories/menus/components/Menu.stories.tsx
@@ -1,5 +1,4 @@
 import { Meta, Story } from '@storybook/react';
-import React from 'react';
 
 import { Box } from '../../../src/Box';
 import { Menu, MenuProps } from '../../../src/Menus/components/Menu';